requirements.txt
bcrypt==4.3.0
click==8.1.8
colorama==0.4.6
fastapi==0.103.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.2.0
PyJWT==2.10.1
python-multipart==0.0.6
pydantic==2.4.2
pydantic-core==2.10.1
starlette==0.27.0
typing-extensions==4.8.0
uvicorn==0.23.2
